添加重定向功能
