+++ src/main/java/frodez/service/user/facade/ILoginService.java
@@ -1,8 +1,10 @@
 package frodez.service.user.facade;
 
 import frodez.config.aop.validation.annotation.Check;
-import frodez.dao.param.user.DoLogin;
-import frodez.dao.param.user.DoRefresh;
+import frodez.config.swagger.annotation.Success;
+import frodez.dao.param.login.DoLogin;
+import frodez.dao.param.login.DoRefresh;
+import frodez.dao.result.login.RefreshInfo;
 import frodez.util.beans.result.Result;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
@@ -20,6 +22,7 @@ public interface ILoginService {
 	 * @date 2018-12-03
 	 */
 	@Check
+	@Success(String.class)
 	Result login(@Valid @NotNull DoLogin param);
 
 	/**
@@ -28,6 +31,7 @@ public interface ILoginService {
 	 * @date 2019-02-27
 	 */
 	@Check
+	@Success(RefreshInfo.class)
 	Result refresh(@Valid @NotNull DoRefresh param);
 
 	/**
